+++ lib/gitlab.dart
@@ -20,14 +20,17 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:gitlab/src/http_client.dart';
+import 'package:gitlab/src/json_map.ext.dart';
 import 'package:intl/intl.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 
 part 'src/commits.dart';
+part 'src/discussions.dart';
 part 'src/issues.dart';
 part 'src/jobs.dart';
 part 'src/merge_requests.dart';
+part 'src/notes.dart';
 part 'src/pipelines.dart';
 part 'src/projects.dart';
 part 'src/snippets.dart';
@@ -35,7 +38,7 @@ part 'src/utils.dart';
 
 final _log = new Logger('GitLab');
 
-enum HttpMethod { get, post, put }
+enum HttpMethod { get, post, put, delete }
 
 /// The main class and entry point to use this library.
 ///
@@ -46,15 +49,36 @@ class GitLab {
   final String host;
   final String scheme;
 
+  /// Assume utf8 within in response bodies.
+  ///
+  /// The response of current gitlab instances (checked with 12.7.0-pre)
+  /// respond without specifying the encoding of the response within the
+  /// content-type (the server responds with content-type: application/json).
+  ///
+  /// The default implementation of http Response does assume latin1 in such
+  /// cases. Therefore any special characters are broken.
+  ///
+  /// In order to avoid this behavior, the utf8 charset, which actually is used
+  /// by gitlab, will be appended to the content-type, thus the resulting header
+  /// is content-type: application/json; charset=utf-8 and will be parsed
+  /// correctly.
+  ///
+  /// This behavior is enabled by default, but one can disable it, because this
+  /// field is introduced in a later version (v0.5.0) and might interfere with
+  /// existing implementations.
+  ///
+  final bool assumeUtf8;
+
   final GitLabHttpClient _httpClient;
 
   static const String apiVersion = 'v4';
 
-  GitLab(this.token, {this.host: 'gitlab.com', this.scheme: 'https'})
+  GitLab(this.token,
+      {this.host = 'gitlab.com', this.scheme = 'https', this.assumeUtf8 = true})
       : _httpClient = new GitLabHttpClient();
 
   GitLab._test(GitLabHttpClient httpClient, this.token,
-      {this.host: 'gitlab.com', this.scheme: 'https'})
+      {this.host: 'gitlab.com', this.scheme: 'https', this.assumeUtf8 = true})
       : _httpClient = httpClient;
 
   /// Get the [ProjectsApi] for this [id].
@@ -77,6 +101,13 @@ class GitLab {
 
     if (!(response.statusCode >= 200 && response.statusCode < 300)) {
       throw new GitLabException(response.statusCode, response.body);
+    }
+
+    final contentType = response.headers["content-type"];
+    final hasContentType = contentType != null;
+    final hasCharset = contentType?.contains("charset") ?? false;
+    if (assumeUtf8 && hasContentType && !hasCharset) {
+      response.headers["content-type"] = "$contentType; charset=utf-8";
     }
 
     return asJson ? jsonDecode(response.body) : response.body;